Added a new file, PluralsightCourses.md, to track the various Pluralsight courses watched and those to watch.
Note that this file also contains the date when the course was watched - Many courses are watched over a number of days/weeks, so the "watched" date represents the date that the course was completed.  Also note this is not currently complete and doesn't contain many courses that have already been watched.